refactor transaction Method
